Removed testing api, edit in api route, post HTTP to database
Removed /<name> where the server returns Hello <name> back to the client. Found it unnecessary after setting up it right. It was developed at the start of the project for testing purposes which we have surpassed. Created firebase class in python as a singleton to make sure the class is only init once, init after will refer to the origin class. This is to prevent multiple firebase to run at the same time. Made Also made it OOP... which made me realize something about it. When creating methods in python class, the first argument is to the class itself. Arguments after is input.